Extracting sufia generator behavior
The generators for sufia needed to be split apart with those relevant
for sufia-models moved to sufia-models and invoked via sufia's install
process.
