--- lib/generators/sufia/templates/config/mailboxer.rb
@@ -1,31 +1,0 @@
-# Copyright © 2012 The Pennsylvania State University
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-# http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-Mailboxer.setup do |config|
-  
-  #Configures if you applications uses or no the email sending for Notifications and Messages
-  config.uses_emails = true
-  
-  #Configures the default from for the email sent for Messages and Notifications of Mailboxer
-  config.default_from = "no-reply@mailboxer.com"
-  
-  #Configures the methods needed by mailboxer
-  config.email_method = :mailboxer_email
-  config.name_method = :name
-
-  #Configures if you use or not a search engine and wich one are you using
-  #Supported enignes: [:solr,:sphinx] 
-  config.search_enabled = false
-  config.search_engine = :solr
-end
