+++ sufia-models/lib/generators/sufia/models/templates/config/clamav.rb
@@ -1,0 +1,1 @@
+ClamAV.instance.loaddb() if defined? ClamAV
