+++ sufia-models/lib/generators/sufia/models/templates/config/redis.yml
