+++ sufia-models/lib/generators/sufia/models/templates/config/resque_config.rb
