+++ sufia-models/lib/generators/sufia/models/templates/config/setup_mail.rb
@@ -1,0 +1,3 @@
+ActionMailer::Base.smtp_settings = {
+  :address                  => "smtp.psu.edu"
+}
